cr_autos/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.4.0
django-types==0.19.1
djangorestframework==3.15.1
sqlparse==0.5.0
types-psycopg2==2.9.21.20240417
